+++ src/app/dialog/dialog-settings.ts
@@ -1,0 +1,23 @@
+const dialogSettingsJoin: any = {
+  dialogText: 'Joining game',
+  showSpinner: true,
+  showConfirm: false,
+};
+
+const dialogSettingsStart: any = {
+  dialogText: 'Starting game',
+  showSpinner: true,
+  showConfirm: false,
+};
+
+const dialogSettingsCreate: any = {
+  dialogText: 'Creating game',
+  showSpinner: true,
+  showConfirm: false,
+};
+
+export const DialogSettings: any = {
+  dialogSettingsCreate: dialogSettingsCreate,
+  dialogSettingsJoin: dialogSettingsJoin,
+  dialogSettingsStart: dialogSettingsStart,
+};