+++ src/main.ts
@@ -13,11 +13,12 @@ const routes: Routes = [
   { path: 'lobby', component: LobbyComponent },
   { path: 'versus-menu', component: VersusMenuComponent },
   { path: 'game', component: MainMenuComponent }, //deprecated, now goes to main menu
+  { path: 'solo', component: MainMenuComponent }, //deprecated, now goes to main menu
   { path: 'versus', component: MainMenuComponent },
   { path: 'join', component: LobbyComponent },
   { path: 'join/:gameCode', component: LobbyComponent },
-  { path: 'solo', component: GameComponent },
-  { path: 'solo/:gameSeed', component: GameComponent },
+  { path: 'endless', component: GameComponent },
+  { path: 'endless/:gameSeed', component: GameComponent },
   { path: 'challenge/:gameSeed', component: MainMenuComponent },
   { path: 'versus/:gameSeed', component: GameComponent },
   { path: 'daily', component: MainMenuComponent },